Move tools subpackage into new repository
